# docs/api/openapi-v1.snapshot.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Agentor.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/ops/queue": {
      "get": {
        "tags": [
          "System"
        ],
        "summary": "Read-only operational queue status for recent queued runs.",
        "operationId": "GetOpsQueue",
        "parameters": [
          {
            "name": "take",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/ops/outbox": {
      "get": {
        "tags": [
          "System"
        ],
        "summary": "Read-only operational status for outbox messages.",
        "operationId": "GetOpsOutbox",
        "parameters": [
          {
            "name": "take",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/ops/leases": {
      "get": {
        "tags": [
          "System"
        ],
        "summary": "Read-only operational status for active execution leases.",
        "operationId": "GetOpsLeases",
        "parameters": [
          {
            "name": "take",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/ops/diagnostics-report": {
      "get": {
        "tags": [
          "System"
        ],
        "summary": "Redacted operator diagnostics bundle (JSON default; format=markdown for Markdown).",
        "operationId": "GetOpsDiagnosticsReport",
        "parameters": [
          {
            "name": "format",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
...
}